add 1st proposal of Apx gui
I do not like it, I find that the workflow is confusing and complex, since the user has to keep open and close views to perform changes.

Since subsystems, stacks, and package managers are related, it will be usual for the user to switch from one to another. Having to open a subsystem, then exit it to open a stack, then exit it to open a package manager... is time-wasting.

The 2nd proposal I am working on is using an IDE-like approach (GNOME Builder inspired). It involves a general add button that shows a menu to choose between creating a new subsystem, stack, or package manager. A sidebar on the left will organize all the subsystems, stacks, and package managers in an AdwViewStack. On the right, there will be a viewport with tabs for each opened resource. Each tab will display the name of the opened resource and a dedicated icon per category (subsystem, stack, package manager) to better identify them. In addition, there will be a headerbar at the top that shows actions based on the opened tab. For example, play/stop and the status if a subsystem, export/import if a stack or package manager, and other relevant features as needed. This way, the user will be able to open as many resources as they need and close them when they are no longer useful.